refactor serde for RegExpReplace
