+++ spark/src/main/scala/org/apache/comet/serde/strings.scala
@@ -21,7 +21,7 @@ package org.apache.comet.serde
 
 import java.util.Locale
 
-import org.apache.spark.sql.catalyst.expressions.{Attribute, Cast, Expression, InitCap, Like, Literal, Lower, RLike, StringLPad, StringRepeat, StringRPad, Substring, Upper}
+import org.apache.spark.sql.catalyst.expressions.{Attribute, Cast, Expression, InitCap, Like, Literal, Lower, RegExpReplace, RLike, StringLPad, StringRepeat, StringRPad, Substring, Upper}
 import org.apache.spark.sql.types.{DataTypes, LongType, StringType}
 
 import org.apache.comet.CometConf
@@ -194,6 +194,49 @@ object CometStringLPad extends CometExpressionSerde[StringLPad] {
       exprToProtoInternal(expr.str, inputs, binding),
       exprToProtoInternal(expr.len, inputs, binding),
       exprToProtoInternal(expr.pad, inputs, binding))
+  }
+}
+
+object CometRegExpReplace extends CometExpressionSerde[RegExpReplace] {
+  override def getSupportLevel(expr: RegExpReplace): SupportLevel = {
+    if (!RegExp.isSupportedPattern(expr.regexp.toString) &&
+      !CometConf.COMET_REGEXP_ALLOW_INCOMPATIBLE.get()) {
+      withInfo(
+        expr,
+        s"Regexp pattern ${expr.regexp} is not compatible with Spark. " +
+          s"Set ${CometConf.COMET_REGEXP_ALLOW_INCOMPATIBLE.key}=true " +
+          "to allow it anyway.")
+      return Incompatible()
+    }
+    expr.pos match {
+      case Literal(value, DataTypes.IntegerType) if value == 1 => Compatible()
+      case _ =>
+        Unsupported(Some("Comet only supports regexp_replace with an offset of 1 (no offset)."))
+    }
+  }
+
+  override def convert(
+      expr: RegExpReplace,
+      inputs: Seq[Attribute],
+      binding: Boolean): Option[Expr] = {
+    expr.pos match {
+      case Literal(value, DataTypes.IntegerType) if value == 1 =>
+        val subjectExpr = exprToProtoInternal(expr.subject, inputs, binding)
+        val patternExpr = exprToProtoInternal(expr.regexp, inputs, binding)
+        val replacementExpr = exprToProtoInternal(expr.rep, inputs, binding)
+        // DataFusion's regexp_replace stops at the first match. We need to add the 'g' flag
+        // to apply the regex globally to match Spark behavior.
+        val flagsExpr = exprToProtoInternal(Literal("g"), inputs, binding)
+        val optExpr = scalarFunctionExprToProto(
+          "regexp_replace",
+          subjectExpr,
+          patternExpr,
+          replacementExpr,
+          flagsExpr)
+        optExprWithInfo(optExpr, expr, expr.subject, expr.regexp, expr.rep, expr.pos)
+      case _ =>
+        None
+    }
   }
 }
